# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Elaniin",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/products/": {
            "get": {
                "tags": [
                    "Products End points"
                ],
                "summary": "Get the list of products",
                "operationId": "c98ca8808a2163d50d2dab3d7062a432",
                "responses": {
                    "200": {
                        "description": "Return a list of products"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/products/10": {
            "get": {
                "tags": [
                    "Products End points"
                ],
                "summary": "Get One product by id",
                "operationId": "0f0f7b99813f6e997e3be18aad8d41cc",
                "responses": {
                    "200": {
                        "description": "Return one product"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/products/search": {
            "post": {
                "tags": [
                    "Products End points"
                ],
                "summary": "search product by Product Name, Name and SKU",
                "operationId": "42bad88497a6972707ac7672fd5eb6e3",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "product Name"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "SKU of product"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return one product"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/products/store": {
            "post": {
                "tags": [
                    "Products End points"
                ],
                "summary": "Save Product",
                "operationId": "90841008e6d2f8cb3655547d2377fb01",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "amount",
                        "required": true,
                        "example": "20"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "price",
                        "required": true,
                        "example": "40.69"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description",
                        "example": "loremp ipsum"
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "image",
                        "required": true,
                        "example": "https://via.placeholder.com/640x480.png/000022?text=quis"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data save successfully"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Oops something went wrong."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/products/update/90": {
            "put": {
                "tags": [
                    "Products End points"
                ],
                "summary": "Update Product",
                "operationId": "8e2282e0038a98071f51f5d7384a1fc6",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "amount",
                        "required": true,
                        "example": "20"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "price",
                        "required": true,
                        "example": "20"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description",
                        "example": "loremp ipsum"
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "image",
                        "required": true,
                        "example": "https://via.placeholder.com/640x480.png/000022?text=quis"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data save successfully"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Oops something went wrong."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/products/7": {
            "delete": {
                "tags": [
                    "Products End points"
                ],
                "summary": "Soft Delete One product by id",
                "operationId": "fd4d33671823c9bbf79d167dcbb57bc1",
                "responses": {
                    "204": {
                        "description": "Data deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}